dgproject/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
Django==5.1.3
djangorestframework==3.15.2
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.9.0
filelock==3.16.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.8
langchain-text-splitters==0.3.2
langsmith==0.1.143
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==1.54.4
orjson==3.10.11
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.2.0
psutil==6.0.0
pure-eval==0.2.2
pydantic==2.8.2
pydantic-settings==2.6.1
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
sentence-transformers==3.3.0
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.2
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.5.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.46.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Wikipedia-API==0.6.0
yarl==1.17.2
zipp==3.19.2
